commenting, cleaning up
