+++ app/src/main/AndroidManifest.xml
@@ -15,16 +15,12 @@
         android:theme="@style/Theme.QuizGo"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
-        <activity
-            android:name=".Activities.MenuActivity"
-            android:exported="true">
-        </activity>
 
         <activity
             android:name=".Activities.GameActivity"
             android:exported="false" />
         <activity
-            android:name=".Activities.UserActivity"
+            android:name=".Activities.MenuActivity"
             android:exported="true">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
