+++ app/src/main/java/hi/hbv601g/QuizGo/Services/GameService.java
@@ -21,7 +21,6 @@ import java.util.List;
 import java.util.concurrent.Semaphore;
 
 import hi.hbv601g.QuizGo.Activities.MenuActivity;
-import hi.hbv601g.QuizGo.Activities.UserActivity;
 import hi.hbv601g.QuizGo.Entities.Question;
 import hi.hbv601g.QuizGo.Entities.User;
 
@@ -35,7 +34,7 @@ public class GameService extends Service {
 
     public GameService() {
         mCurrentPlayer = 0;
-        mUsers = UserActivity.getUserService().getUsers();
+        mUsers = MenuActivity.getUserService().getUsers();
 
         //semaphore for question api-call
         sem = new Semaphore(0);
