--- app/src/main/res/layout/activity_user.xml
@@ -1,148 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#FF9A16"
-    android:gravity="center"
-    android:orientation="vertical"
-    tools:context="hi.hbv601g.QuizGo.Activities.UserActivity">
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="45dp" />
-
-    <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="match_parent"
-        android:layout_height="165dp"
-        app:srcCompat="@drawable/logo" />
-
-    <EditText
-        android:id="@+id/username"
-        android:layout_width="254dp"
-        android:layout_height="48dp"
-        android:ems="10"
-        android:hint="@string/username"
-        android:inputType="textPersonName"
-        android:textAlignment="center" />
-
-    <EditText
-        android:id="@+id/password"
-        android:layout_width="256dp"
-        android:layout_height="48dp"
-        android:ems="10"
-        android:hint="@string/password"
-        android:inputType="textPassword"
-        android:textAlignment="center" />
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="28dp" />
-
-    <Button
-        android:id="@+id/loginButton"
-        style="@style/Widget.AppCompat.Button.Colored"
-        android:layout_width="256dp"
-        android:layout_height="wrap_content"
-        android:allowUndo="false"
-        android:text="@string/login"
-        android:textColor="#1824FC"
-        android:textSize="20sp"
-        app:backgroundTint="#FF9A16" />
-
-    <Button
-        android:id="@+id/registerButton"
-        style="@style/Widget.AppCompat.Button.Colored"
-        android:layout_width="256dp"
-        android:layout_height="wrap_content"
-        android:text="@string/register"
-        android:textColor="#1824FC"
-        android:textSize="20sp"
-        app:backgroundTint="#FF9A16" />
-
-
-    <Button
-        android:id="@+id/logoutButton"
-        style="@style/Widget.AppCompat.Button.Colored"
-        android:layout_width="256dp"
-        android:layout_height="wrap_content"
-        android:allowUndo="false"
-        android:text="@string/logout"
-        android:textColor="#1824FC"
-        android:textSize="20sp"
-        app:backgroundTint="#FF9A16" />
-
-    <Button
-        android:id="@+id/playButton"
-        style="@style/Widget.AppCompat.Button.Colored"
-        android:layout_width="256dp"
-        android:layout_height="wrap_content"
-        android:text="@string/playGame"
-        android:textColor="#1824FC"
-        android:textSize="20sp"
-        app:backgroundTint="#FF9A16" />
-
-
-    <Button
-        android:id="@+id/testButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Testing" />
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="26dp" />
-
-    <LinearLayout
-        android:layout_width="340dp"
-        android:layout_height="75dp"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/user1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textAlignment="center"
-            tools:text="User1" />
-
-        <TextView
-            android:id="@+id/user2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textAlignment="center"
-            tools:text="User2" />
-
-        <TextView
-            android:id="@+id/user3"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textAlignment="center"
-            tools:text="User3" />
-
-        <TextView
-            android:id="@+id/user4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textAlignment="center"
-            tools:text="User4" />
-
-    </LinearLayout>
-
-    <RadioGroup
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Button" />
-
-</LinearLayout>
